# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de controle financeiro pessoal com autenticação JWT",
        "title": "Imperium API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/profile/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna estatísticas financeiras do usuário para exibição em gráficos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Obter estatísticas do perfil",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProfileStats"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/profile/stats/breakdown": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna despesas detalhadas por categoria com sub-categorias",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Obter breakdown de despesas",
                "parameters": [
                    {
                        "type": "string",
                        "default": "month",
                        "description": "Período: 'week', 'month', 'year', 'all'",
                        "name": "period",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/profile/stats/daily": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna dados de gastos/receitas por dia para gráfico de barras",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Obter estatísticas diárias",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 30,
                        "description": "Número de dias",
                        "name": "days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/profile/stats/income": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna receitas detalhadas por categoria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Obter breakdown de receitas",
                "parameters": [
                    {
                        "type": "string",
                        "default": "month",
                        "description": "Período: 'week', 'month', 'year', 'all'",
                        "name": "period",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CategoryStat": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "percentage": {
                    "type": "number"
                }
            }
        },
        "models.ComparisonStats": {
            "type": "object",
            "properties": {
                "expense_change": {
                    "description": "Percentage change",
                    "type": "number"
                },
                "income_change": {
                    "description": "Percentage change",
                    "type": "number"
                },
                "savings_change": {
                    "description": "Percentage change",
                    "type": "number"
                }
            }
        },
...
        "models.MonthlyTrendPoint": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "expenses": {
                    "type": "number"
                },
                "income": {
                    "type": "number"
                },
                "month": {
                    "description": "\"2024-01\", \"2024-02\", etc",
                    "type": "string"
                }
            }
        },
...
        "models.ProfileStats": {
            "type": "object",
            "properties": {
                "comparison_last_month": {
                    "$ref": "#/definitions/models.ComparisonStats"
                },
                "connected_accounts": {
                    "type": "integer"
                },
                "expenses_by_category": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CategoryStat"
                    }
                },
                "monthly_expenses": {
                    "type": "number"
                },
                "monthly_income": {
                    "type": "number"
                },
                "monthly_savings": {
                    "type": "number"
                },
                "monthly_trend": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MonthlyTrendPoint"
                    }
                },
                "top_categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CategoryStat"
                    }
                },
                "total_balance": {
                    "type": "number"
                },
                "transaction_count": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite: Bearer {seu_token_aqui}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}